LPD-53900 Moves classes no logic changes
