+++ modules/apps/style-book/style-book-item-selector-api/src/main/java/com/liferay/style/book/item/selector/criterion/StyleBookEntryItemSelectorCriterion.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.admin.web.internal.item.selector;
+package com.liferay.style.book.item.selector.criterion;
 
 import com.liferay.item.selector.BaseItemSelectorCriterion;
