+++ modules/apps/style-book/style-book-item-selector-web/src/main/java/com/liferay/style/book/item/selector/web/internal/StyleBookEntryItemDescriptor.java
@@ -3,15 +3,15 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.admin.web.internal.item.selector;
+package com.liferay.style.book.item.selector.web.internal;
 
 import com.liferay.frontend.taglib.clay.servlet.taglib.VerticalCard;
 import com.liferay.item.selector.ItemSelectorViewDescriptor;
-import com.liferay.layout.admin.web.internal.frontend.taglib.clay.servlet.taglib.SelectStylebookLayoutVerticalCard;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.dao.search.RowChecker;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.model.Layout;
+import com.liferay.style.book.item.selector.web.internal.frontend.taglib.clay.servlet.taglib.SelectStylebookLayoutVerticalCard;
 import com.liferay.style.book.model.StyleBookEntry;
 
 import java.util.Locale;
