+++ modules/apps/style-book/style-book-item-selector-web/src/main/java/com/liferay/style/book/item/selector/web/internal/StyleBookEntryItemSelectorCriterionHandler.java
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.admin.web.internal.item.selector;
+package com.liferay.style.book.item.selector.web.internal;
 
 import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
+import com.liferay.style.book.item.selector.criterion.StyleBookEntryItemSelectorCriterion;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
