+++ modules/apps/style-book/style-book-item-selector-web/src/main/java/com/liferay/style/book/item/selector/web/internal/frontend/taglib/clay/servlet/taglib/SelectStylebookLayoutVerticalCard.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.admin.web.internal.frontend.taglib.clay.servlet.taglib;
+package com.liferay.style.book.item.selector.web.internal.frontend.taglib.clay.servlet.taglib;
 
 import com.liferay.frontend.taglib.clay.servlet.taglib.VerticalCard;
 import com.liferay.petra.string.StringPool;
